EXPAND with checkbox in same panel/background
